# services/user-service/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: User Service API
  description: |
    This API provides access to user-related functionalities in the IS/STAG system.
    It allows for the management and retrieval of user data.
  version: 1.0.0
  contact:
    name: STAG Support
    email: stag@service.zcu.cz
    url: https://is-stag.zcu.cz/zakaznici/Support.html

servers:
  - url: /api/v1
    description: User Service API through API Gateway

tags:
  - name: persons
    description: Operations for managing persons

paths:

  /persons/{personId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - persons
      summary: Get person
      description: Retrieve a person by their unique identifier.
      operationId: getPersonProfile
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/PersonId'
      responses:
        '200':
          description: Person details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '503':
          $ref: '#/components/responses/503ServiceUnavailable'
        'default':
          $ref: '#/components/responses/Default'
    patch:
      security:
        - bearerAuth: [ ]
      tags:
        - persons
      summary: Update person
      description: Update a person by their unique identifier.
      operationId: updatePersonProfile
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/PersonId'
      requestBody:
        description: Person details to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonRequest'
      responses:
        '204':
          description: Person details updated successfully
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '503':
          $ref: '#/components/responses/503ServiceUnavailable'
        'default':
          $ref: '#/components/responses/Default'

  /persons/{personId}/addresses:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - persons
      summary: Get person addresses
      description: Retrieve addresses associated with a person by their unique identifier.
      operationId: getPersonAddresses
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/PersonId'
      responses:
        '200':
          description: Person addresses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressesResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '503':
          $ref: '#/components/responses/503ServiceUnavailable'
        'default':
          $ref: '#/components/responses/Default'

  /persons/{personId}/banking:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - persons
      summary: Get person banking information
      description: Retrieve banking information associated with a person by their unique identifier.
      operationId: getPersonBanking
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/PersonId'
      responses:
        '200':
          description: Person banking information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountsResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '503':
          $ref: '#/components/responses/503ServiceUnavailable'
        'default':
          $ref: '#/components/responses/Default'

  /persons/{personId}/education:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - persons
      summary: Get person education information
      description: Retrieve education information associated with a person by their unique identifier.
      operationId: getPersonEducation
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/PersonId'
      responses:
        '200':
          description: Person education information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '503':
          $ref: '#/components/responses/503ServiceUnavailable'
        'default':
          $ref: '#/components/responses/Default'

components:

  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Keycloak

  parameters:

    AcceptLanguage:
      name: Accept-Language
      in: header
      description: Preferred language
      required: false
      schema:
        type: string
        enum:
          - cs
          - en
        default: cs

    PersonId:
      name: personId
      in: path
      required: true
      description: Unique identifier of the person.
      schema:
        $ref: '#/components/schemas/PersonId'

  schemas:

    PersonId:
      type: integer
      format: int32
      nullable: false
      title: Person ID
      description: The unique numeric identifier for a person.
      example: 221264
      minimum: 1

    StudentId:
      type: string
      title: Student ID
      nullable: false
      description: The unique alphanumeric identifier for a student.
      maxLength: 10
      example: K24B4302P

    UpdatePersonRequest:
      type: object
      title: Update Person Request
      description: Request containing person details to be updated.
      additionalProperties: false
      nullable: false
      properties:
        birthSurname:
          $ref: '#/components/schemas/PersonResponse/properties/birthSurname'
        maritalStatus:
          $ref: '#/components/schemas/PersonResponse/properties/maritalStatus'
        contact:
          $ref: '#/components/schemas/ContactUpdate'
        titles:
          $ref: '#/components/schemas/TitlesUpdate'
        birthPlace:
          $ref: '#/components/schemas/BirthPlaceUpdate'
        temporaryAddress:
          $ref: '#/components/schemas/AddressUpdate'
        bankAccount:
          $ref: '#/components/schemas/BankAccountUpdate'

    BirthPlaceUpdate:
      type: object
      title: Birth Place Update
      description: Object containing birth place to be updated.
      additionalProperties: false
      nullable: false
      properties:
        country:
          $ref: '#/components/schemas/BirthPlace/properties/country'
        city:
          $ref: '#/components/schemas/BirthPlace/properties/city'

    TitlesUpdate:
      type: object
      title: Titles Update
      description: Object containing titles to be updated.
      additionalProperties: false
      nullable: false
      properties:
        prefix:
          $ref: '#/components/schemas/Title'
        suffix:
          $ref: '#/components/schemas/Title'

    ContactUpdate:
      type: object
      title: Contact Update
      description: Object containing contact details to be updated.
      additionalProperties: false
      nullable: false
      properties:
        email:
          $ref: '#/components/schemas/Contact/properties/email'
        phone:
          $ref: '#/components/schemas/Contact/properties/phone'
        mobile:
          $ref: '#/components/schemas/Contact/properties/mobile'
        dataBox:
          $ref: '#/components/schemas/Contact/properties/dataBox'

    AddressUpdate:
      type: object
      title: Address Update
      description: Object containing address details to be updated.
      additionalProperties: false
      nullable: false
      properties:
        country:
          $ref: '#/components/schemas/Address/properties/country'
        district:
          $ref: '#/components/schemas/Address/properties/district'
        municipality:
          $ref: '#/components/schemas/Address/properties/municipality'
        municipalityPart:
          $ref: '#/components/schemas/Address/properties/municipalityPart'
        street:
          $ref: '#/components/schemas/Address/properties/street'
        streetNumber:
          $ref: '#/components/schemas/Address/properties/streetNumber'
        zipCode:
          $ref: '#/components/schemas/Address/properties/zipCode'
        postOffice:
          $ref: '#/components/schemas/ForeignAddress/properties/postOffice'

    BankAccountUpdate:
      type: object
      title: Bank Account Update
      description: Object containing bank account details to be updated.
      additionalProperties: false
      nullable: false
      properties:
        accountNumberPrefix:
          $ref: '#/components/schemas/BankAccount/properties/accountNumberPrefix'
        accountNumberSuffix:
          $ref: '#/components/schemas/BankAccount/properties/accountNumberSuffix'
        bankCode:
          $ref: '#/components/schemas/BankAccount/properties/bankCode'
        holderName:
          $ref: '#/components/schemas/BankAccount/properties/holderName'
        holderAddress:
          $ref: '#/components/schemas/BankAccount/properties/holderAddress'

    PersonResponse:
      type: object
      title: Person Response
      description: Response containing person details.
      additionalProperties: false
      nullable: false
      required:
        - personId
        - studentIds
        - firstName
        - lastName
        - birthSurname
        - contact
        - titles
        - birthNumber
        - birthDate
        - birthPlace
        - citizenship
        - passportNumber
        - gender
        - maritalStatus
      properties:
        personId:
          $ref: '#/components/schemas/PersonId'
        studentIds:
          type: array
          nullable: false
          description: List of student ids associated with the person.
          items:
            $ref: '#/components/schemas/StudentId'
          example:
            - K24B4302P
        firstName:
          type: string
          nullable: false
          description: First name of the person.
          example: John
          maxLength: 100
        lastName:
          type: string
          nullable: false
          description: Last name of the person.
          example: Doe
          maxLength: 100
        birthSurname:
          title: Birth Surname
          type: string
          nullable: true
          description: Birth surname of the person.
          example: Smith
          maxLength: 100
        contact:
          $ref: '#/components/schemas/Contact'
        titles:
          $ref: '#/components/schemas/Titles'
        birthNumber:
          type: string
          nullable: false
          description: Birth number of the person.
          example: '7061030163'
          pattern: ^\d{10}$
          maxLength: 10
        birthDate:
          type: string
          format: date
          nullable: false
          description: Birth date of the person.
          example: '1990-01-01'
        birthPlace:
          $ref: '#/components/schemas/BirthPlace'
        citizenship:
          $ref: '#/components/schemas/Citizenship'
        passportNumber:
          type: string
          nullable: true
          description: Passport number of the person.
          example: '16128791799'
          pattern: ^[A-Z0-9]{1,11}$
          maxLength: 11
        gender:
          type: string
          nullable: false
          description: Gender of the person.
          example: male
          maxLength: 240
        maritalStatus:
          title: Marital Status
          type: string
          nullable: true
          description: Marital status of the person.
          example: single
          maxLength: 240

    Citizenship:
      type: object
      title: Citizenship
      description: Citizenship details of the person.
      additionalProperties: false
      nullable: false
      required:
        - country
        - qualifier
      properties:
        country:
          type: string
          nullable: false
          description: Country of citizenship.
          example: Czech Republic
          maxLength: 70
        qualifier:
          type: string
          nullable: true
          description: Citizenship qualifier of the person.
          example: Občan
          maxLength: 240

    BirthPlace:
      type: object
      title: Birth Place
      description: Birth place of the person.
      additionalProperties: false
      nullable: false
      required:
        - country
        - city
      properties:
        country:
          title: Country
          type: string
          nullable: false
          description: Birth country of the person.
          example: Czech Republic
          maxLength: 70
        city:
          title: City
          type: string
          nullable: true
          description: Birth city of the person.
          example: Prague
          maxLength: 100

    Titles:
      type: object
      title: Titles
      description: Titles of the person.
      additionalProperties: false
      nullable: false
      required:
        - prefix
        - suffix
      properties:
        prefix:
          $ref: '#/components/schemas/Title'
        suffix:
          $ref: '#/components/schemas/Title'

    Title:
      type: string
      nullable: true
      description: Title of the person.
      example: Ing.
      maxLength: 240

    Contact:
      type: object
      title: Contact
      description: Contact information of the person.
      additionalProperties: false
      nullable: false
      required:
        - email
        - phone
        - mobile
        - dataBox
      properties:
        email:
          title: Email
          type: string
          format: email
          nullable: true
          description: Email address of the person.
          example: Martin95043186@xxx.cz
        phone:
          title: Phone
          type: string
          nullable: true
          description: Phone number of the person.
          example: '+420123456789'
          pattern: ^\+?\d{1,20}$
          maxLength: 20
        mobile:
          title: Mobile
          type: string
          nullable: true
          description: Mobile number of the person.
          example: '+420987654321'
          pattern: ^\+?\d{1,30}$
          maxLength: 30
        dataBox:
          title: Data Box
          type: string
          nullable: true
          description: Data box number of the person.
          example: 'krmm7m5'
          pattern: ^[a-km-np-z2-9]{7}$
          maxLength: 7

    Address:
      type: object
      title: Address
      description: Address details of the person.
      additionalProperties: false
      nullable: false
      required:
        - street
        - streetNumber
        - zipCode
        - municipality
        - municipalityPart
        - district
        - country
      properties:
        street:
          title: Street
          type: string
          nullable: true
          description: Street name of the address.
          example: Příchovice
          maxLength: 48
        streetNumber:
          title: Street Number
          type: string
          nullable: true
          description: Street number of the address.
          example: '97'
          maxLength: 10
        zipCode:
          title: Zip Code
          type: string
          nullable: true
          description: Zip code of the address.
          example: '34901'
          pattern: ^\d{1,5}$
          maxLength: 5
        municipality:
          title: Municipality
          type: string
          nullable: true
          description: Municipality of the address.
          example: Stříbro
          maxLength: 255
        municipalityPart:
          title: Municipality Part
          type: string
          nullable: true
          description: Part of the municipality.
          example: Stříbro I
          maxLength: 255
        district:
          title: District
          type: string
          nullable: true
          description: District of the address.
          example: Tachov
          maxLength: 100
        country:
          title: Country
          type: string
          nullable: true
          description: Country of the address.
          example: Česká republika
          maxLength: 70

    ForeignAddress:
      type: object
      title: Foreign Address
      description: Address details of the person when they have a foreign residence.
      additionalProperties: false
      nullable: false
      required:
        - zipCode
        - municipality
        - district
        - postOffice
      properties:
        zipCode:
          title: Zip Code
          type: string
          nullable: true
          description: Zip code of the foreign address.
          example: '1110'
          pattern: ^\d{1,10}$
          maxLength: 10
        municipality:
          title: Municipality
          type: string
          nullable: true
          description: Municipality of the foreign address.
          example: Mornaguia
          maxLength: 75
        district:
          title: District
          type: string
          nullable: true
          description: District of the foreign address.
          example: manouba
          maxLength: 50
        postOffice:
          title: Post Office
          type: string
          nullable: true
          description: Post office of the foreign address.
          example: Mornaguia
          maxLength: 75

    AddressesResponse:
      type: object
      title: Person Addresses Response
      description: Response containing person addresses.
      additionalProperties: false
      nullable: false
      required:
        - permanentAddress
        - temporaryAddress
        - foreignPermanentAddress
        - foreignTemporaryAddress
      properties:
        permanentAddress:
          $ref: '#/components/schemas/Address'
        temporaryAddress:
          $ref: '#/components/schemas/Address'
        foreignPermanentAddress:
          $ref: '#/components/schemas/ForeignAddress'
        foreignTemporaryAddress:
          $ref: '#/components/schemas/ForeignAddress'

    BankAccount:
      type: object
      title: Bank Account
      description: Response containing bank account details.
      nullable: false
      required:
        - accountNumberPrefix
        - accountNumberSuffix
        - bankCode
        - bankName
        - iban
        - currency
        - holderName
        - holderAddress
      properties:
        accountNumberPrefix:
          title: Account Number Prefix
          type: string
          nullable: true
          description: Prefix of the bank account number.
          example: "295538"
          pattern: ^\d{1,6}$
          minLength: 1
          maxLength: 6
        accountNumberSuffix:
          title: Account Number Suffix
          type: string
          nullable: true
          description: Main part of the bank account number.
          example: "4167197191"
          pattern: ^\d{1,10}$
          minLength: 1
          maxLength: 10
        bankCode:
          title: Bank Code
          type: string
          nullable: true
          description: Bank code of the bank where the account is held.
          example: "0100"
          pattern: ^\d{4}$
          maxLength: 4
        bankName:
          title: Bank Name
          type: string
          nullable: true
          description: Name of the bank where the account is held.
          example: Česká spořitelna, a.s.
          maxLength: 240
        iban:
          title: IBAN
          type: string
          nullable: true
          description: International Bank Account Number (IBAN).
          example: CZ6508000000002000145399
          maxLength: 30
        currency:
          title: Currency
          type: string
          nullable: true
          description: Currency of the bank account.
          example: CZK
          pattern: ^[A-Z]{3}$
          maxLength: 3
        holderName:
          title: Holder Name
          type: string
          nullable: true
          description: Name of the account holder.
          example: Jan Novák
          maxLength: 255
        holderAddress:
          title: Holder Address
          type: string
          nullable: true
          description: Address of the account holder.
          example: Příchovice 97, 34901 Stříbro, Česká republika
          maxLength: 255

    EuroBankAccount:
      type: object
      title: Euro Bank Account
      description: Bank account details for Euro transactions.
      nullable: false
      allOf:
        - $ref: '#/components/schemas/BankAccount'
        - $ref: '#/components/schemas/CountryAndSwiftCode'
      example:
        accountNumberPrefix: "295538"
        accountNumberSuffix: "4167197191"
        bankCode: "0100"
        bankName: "Česká spořitelna, a.s."
        iban: "CZ6508000000002000145399"
        currency: "EUR"
        holderName: "Jan Novák"
        holderAddress: "Příchovice 97, 34901 Stříbro, Česká republika"
        country: "Česká republika"
        swiftCode: "KOMBCZPPXXX"

    CountryAndSwiftCode:
      type: object
      title: Country And Swift Code
      description: Country and SWIFT code of the bank for Euro transactions.
      nullable: false
      required:
        - country
        - swiftCode
      properties:
        country:
          type: string
          nullable: true
          description: Country code of the Euro account.
          example: Česká republika
          maxLength: 70
        swiftCode:
          type: string
          nullable: true
          description: SWIFT code of the bank for Euro transactions.
          example: KOMBCZPPXXX
          maxLength: 11

    BankAccountsResponse:
      type: object
      title: Bank Accounts Response
      description: Bank accounts of the person.
      additionalProperties: false
      nullable: false
      required:
        - account
        - euroAccount
      properties:
        account:
          $ref: '#/components/schemas/BankAccount'
        euroAccount:
          $ref: '#/components/schemas/EuroBankAccount'

    HighSchoolAddress:
      type: object
      title: High School Address
      description: High School address details.
      additionalProperties: false
      nullable: false
      required:
        - street
        - zipCode
        - municipality
        - district
        - country
      properties:
        street:
          type: string
          nullable: true
          description: Street name of the high school address.
          example: Příchovice
          maxLength: 48
        zipCode:
          type: string
          nullable: true
          description: Zip code of the high school address.
          example: '34901'
          pattern: ^\d{1,5}$
          maxLength: 5
        municipality:
          type: string
          nullable: true
          description: Municipality of the high school address.
          example: Stříbro
          maxLength: 255
        district:
          type: string
          nullable: true
          description: District of the high school address.
          example: Tachov
          maxLength: 100
        country:
          type: string
          nullable: true
          description: Country of the high school address.
          example: Česká republika
          maxLength: 70

    EducationResponse:
      type: object
      title: Education Response
      description: Education details of the person.
      additionalProperties: false
      nullable: false
      required:
        - highSchool
        - foreignHighSchool
      properties:
        highSchool:
          $ref: '#/components/schemas/HighSchool'
        foreignHighSchool:
          $ref: '#/components/schemas/ForeignHighSchool'

    HighSchool:
      type: object
      title: High School
      description: Details of the high school attended by the person.
      additionalProperties: false
      nullable: false
      required:
        - schoolName
        - fieldOfStudy
        - graduationDate
        - address
      properties:
        schoolName:
          type: string
          nullable: true
          description: Name of the high school.
          example: Gymnázium
          maxLength: 150
        fieldOfStudy:
          type: string
          nullable: true
          description: Field of study at the high school.
          example: Výpočetní technika a informatika
          maxLength: 240
        graduationDate:
          type: string
          format: date
          nullable: true
          description: Date of graduation from the educational institution.
          example: '2020-06-15'
        address:
          $ref: '#/components/schemas/HighSchoolAddress'

    ForeignHighSchool:
      type: object
      title: Foreign High School
      description: Details of the foreign high school attended by the person.
      additionalProperties: false
      nullable: false
      required:
        - schoolName
        - fieldOfStudy
        - location
      properties:
        schoolName:
          type: string
          nullable: true
          description: Name of the foreign high school.
          example: International School
          maxLength: 75
        fieldOfStudy:
          type: string
          nullable: true
          description: Field of study at the foreign high school.
          example: Computer Science
          maxLength: 255
        location:
          type: string
          nullable: true
          description: Location of the foreign high school.
          example: Paris
          maxLength: 75

    ProblemDetail:
      type: object
      title: Problem Detail
      description: A generic error, as defined by RFC 7807. It can be extended with additional properties.
      additionalProperties: true
      nullable: false
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: http://localhost:8080/person-not-found
          pattern: ^((about:blank)|(https?://[^\s]+))$
          maxLength: 1000
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Domain Not Found
          pattern: ^[A-Za-z0-9\s\-.,!?:;'"]+$
          maxLength: 255
        status:
          type: integer
          format: int32
          description: The HTTP status code.
          example: 404
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: 'Person with ID: 221264 not found'
          pattern: ^[A-Za-z0-9\s\-.,!?:;'"]+$
          maxLength: 1000
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem.
          example: /persons/221264
          pattern: ^(/[^\s]*)+$
          maxLength: 1000

  responses:

    400BadRequest:
      description: Bad request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: http://localhost:8080/api/v1/bad-request
            title: Bad Request
            status: 400
            detail: Invalid person ID
            instance: /api/v1/persons/-10

    401Unauthorized:
      description: Access token is missing or invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: http://localhost:8080/api/v1/unauthorized
            title: Unauthorized
            status: 401
            detail: Access token is missing or invalid
            instance: /api/v1/persons/221264

    403Forbidden:
      description: You do not have permission to access this resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: http://localhost:8080/api/v1/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to access this resource
            instance: /api/v1/persons/221264

    404NotFound:
      description: The specified resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: http://localhost:8080/api/v1/person-not-found
            title: Person Not Found
            status: 404
            detail: 'Person with ID: 221264 not found'
            instance: /api/v1/persons/221264

    500InternalServerError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: http://localhost:8080/api/v1/internal-server-error
            title: Internal server error
            status: 500
            detail: An unexpected error occurred
            instance: /api/v1/persons/221264

    503ServiceUnavailable:
      description: The service is currently unavailable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: http://localhost:8080/api/v1/service-unavailable
            title: Service Unavailable
            status: 503
            detail: Service is currently unavailable. Please try again later.
            instance: /api/v1/persons/221264

    Default:
      description: Unexpected error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'